+++ lib/data_services_api/dataset.rb
@@ -26,6 +26,10 @@ module DataServicesApi
       @json['data-api']
     end
 
+    def dataset
+      @json['dataset']
+    end
+
     def structure_api
       @json['structure-api']
     end
@@ -49,7 +53,10 @@ module DataServicesApi
     end
 
     def query(query)
-      service.api_post_json(data_api, query.to_json)
+      sapi_query_params = SapiNTConverter.new(query.to_json).to_sapint_query
+      sapint_response = service.api_get_json(data_api, sapi_query_params)
+      json_mode_compact = query.terms['@json_mode'] == 'compact'
+      DSAPIResponseConverter.new(sapint_response, dataset, json_mode_compact).to_dsapi_response
     end
 
     def describe(uri)
@@ -57,7 +64,9 @@ module DataServicesApi
     end
 
     def explain(query)
-      service.api_post_json(explain_api, query.to_json)
+      sapi_query_params = SapiNTConverter.new(query.to_json).to_sapint_query
+      explain_url = "#{data_api}/explain"
+      service.api_get_json(explain_url, sapi_query_params)
     end
   end
 end