+++ spec/data_services_api/data_services_api_spec.rb
@@ -7,11 +7,6 @@ describe 'DataServicesAPI', 'the data services API' do
     _(DataServicesApi::VERSION).must_match(/\d+\.\d+\.\d+/)
   end
 
-  it 'should be constructable with a default URL' do
-    dsapi = DataServicesApi::Service .new
-    _(dsapi.url).must_match(%r{localhost:8080/dsapi})
-  end
-
   it 'should be constructable with a given URL' do
     dsapi = DataServicesApi::Service .new(url: 'foo/bar')
     _(dsapi.url).must_match(%r{foo/bar})